exon-py/pyproject.toml
[build-system]
build-backend = "maturin"
requires = ["maturin>=1.1.0"]

[project]
authors = [
  {name = "WHERE TRUE devs", email = "thauck+exonpy@wheretrue.com"},
]
classifiers = [
  "Programming Language :: Rust",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["pyarrow>=12", "adbc-driver-flightsql>=0.6.0", "boto3>=1.28.0", "grpcio>=1.58.0"]

license = {file = "LICENSE"}
name = "exonpy"
readme = "README.md"
requires-python = ">=3.8"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"

[tool.ruff]
exclude = ["python/exonpy/proto", "python/exonpy/proto/*"]
